+++ back/src/main/java/com/ms/back/global/config/SecurityConfig.java
@@ -1,10 +1,10 @@
 package com.ms.back.global.config;
 
-import com.ms.back.member.jwt.CustomLogoutFilter;
-import com.ms.back.member.jwt.JWTFilter;
-import com.ms.back.member.jwt.JWTUtil;
-import com.ms.back.member.jwt.LoginFilter;
-import com.ms.back.member.repository.RefreshRepository;
+import com.ms.back.global.jwt.CustomLogoutFilter;
+import com.ms.back.global.jwt.JWTFilter;
+import com.ms.back.global.jwt.JWTUtil;
+import com.ms.back.global.jwt.LoginFilter;
+import com.ms.back.member.infrastructure.repository.RefreshRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -22,8 +22,6 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 @Configuration
