+++ back/src/main/java/com/ms/back/global/jwt/JWTFilter.java
@@ -1,7 +1,7 @@
-package com.ms.back.member.jwt;
+package com.ms.back.global.jwt;
 
-import com.ms.back.member.dto.CustomUserDetails;
-import com.ms.back.member.entity.UserEntity;
+import com.ms.back.member.application.dto.CustomUserDetails;
+import com.ms.back.member.domain.model.UserEntity;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
