+++ back/src/main/java/com/ms/back/member/application/dto/LoginRequestDTO.java
@@ -1,4 +1,4 @@
-package com.ms.back.member.dto;
+package com.ms.back.member.application.dto;
 
 public class LoginRequestDTO {
     private String username;
