+++ back/src/main/java/com/ms/back/member/application/dto/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ms.back.member.dto;
+package com.ms.back.member.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
