+++ back/src/main/java/com/ms/back/member/application/service/JoinService.java
@@ -1,8 +1,8 @@
-package com.ms.back.member.service;
+package com.ms.back.member.application.service;
 
-import com.ms.back.member.dto.JoinDTO;
-import com.ms.back.member.entity.UserEntity;
-import com.ms.back.member.repository.UserRepository;
+import com.ms.back.member.application.dto.JoinDTO;
+import com.ms.back.member.domain.model.UserEntity;
+import com.ms.back.member.infrastructure.repository.UserRepository;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
