--- back/src/main/java/com/ms/back/member/controller/LoginController.java
@@ -1,12 +1,0 @@
-package com.ms.back.member.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-    @GetMapping("/login")
-    public String loginP(){
-        return "login";
-    }
-}
